+++ RayTracer/src/Camera.h
@@ -1,18 +1,18 @@
 #include "Util.h"
 #include "Ray.h"
 
-#include "vendor/glm/ext/vector_float3.hpp"
+#include "vendor/glm/ext/vector_double3.hpp"
 
 class Camera {
 public:
 	Camera();
 	~Camera();
 
-	Ray getRay(float pixelX, float pixelY) const;
+	Ray getRay(double pixelX, double pixelY) const;
 
 private:
-	glm::vec3 m_Origin;
-	glm::vec3 m_ViewportHorizontal;
-	glm::vec3 m_ViewportVertical;
-	glm::vec3 m_ViewportLowerLeftCorner;
+	glm::dvec3 m_Origin;
+	glm::dvec3 m_ViewportHorizontal;
+	glm::dvec3 m_ViewportVertical;
+	glm::dvec3 m_ViewportLowerLeftCorner;
 };
