+++ RayTracer/src/Ray.cpp
@@ -1,6 +1,6 @@
 #include "Ray.h"
 
-Ray::Ray(const glm::vec3 &origin, const glm::vec3 &direction)
+Ray::Ray(const glm::dvec3 &origin, const glm::dvec3 &direction)
 	: m_Origin(origin), m_Direction(direction)
 {
 
@@ -8,14 +8,14 @@ Ray::Ray(const glm::vec3 &origin, const glm::vec3 &direction)
 
 Ray::~Ray() {}
 
-glm::vec3 Ray::GetOrigin() const {
+glm::dvec3 Ray::GetOrigin() const {
 	return m_Origin;
 }
 
-glm::vec3 Ray::GetDirection() const {
+glm::dvec3 Ray::GetDirection() const {
 	return m_Direction;
 }
 
-glm::vec3 Ray::GetPositionAtDistance(float t) const {
+glm::dvec3 Ray::GetPositionAtDistance(double t) const {
 	return m_Origin + t * m_Direction;
 }
