+++ RayTracer/src/Ray.h
@@ -1,18 +1,18 @@
 #pragma once
 
-#include "vendor/glm/ext/vector_float3.hpp"
+#include "vendor/glm/ext/vector_double3.hpp"
 
 class Ray {
 public:
-	Ray(const glm::vec3 &origin, const glm::vec3 &direction);
+	Ray(const glm::dvec3 &origin, const glm::dvec3 &direction);
 	~Ray();
 
-	glm::vec3 GetOrigin() const;
-	glm::vec3 GetDirection() const;
+	glm::dvec3 GetOrigin() const;
+	glm::dvec3 GetDirection() const;
 
-	glm::vec3 GetPositionAtDistance(float t) const;
+	glm::dvec3 GetPositionAtDistance(double t) const;
 
 private:
-	glm::vec3 m_Origin;
-	glm::vec3 m_Direction;
+	glm::dvec3 m_Origin;
+	glm::dvec3 m_Direction;
 };
