+++ RayTracer/src/Sphere.cpp
@@ -4,23 +4,23 @@
 
 #include "vendor/glm/gtx/norm.hpp"
 
-Sphere::Sphere(glm::vec3 center, float radius)
+Sphere::Sphere(glm::dvec3 center, double radius)
 	: m_Center(center), m_Radius(radius)
 {}
 
-bool Sphere::Hit(const Ray &ray, float tMin, float tMax, HitRecord &hitRecord) const {
+bool Sphere::Hit(const Ray &ray, double tMin, double tMax, HitRecord &hitRecord) const {
 
-    glm::vec3 rayOriginToSphereCenter = ray.GetOrigin() - m_Center;
-    float a = glm::length2(ray.GetDirection());
-    float halfB = glm::dot(rayOriginToSphereCenter, ray.GetDirection());
-    float c = glm::length2(rayOriginToSphereCenter) - m_Radius * m_Radius;
+    glm::dvec3 rayOriginToSphereCenter = ray.GetOrigin() - m_Center;
+    double a = glm::length2(ray.GetDirection());
+    double halfB = glm::dot(rayOriginToSphereCenter, ray.GetDirection());
+    double c = glm::length2(rayOriginToSphereCenter) - m_Radius * m_Radius;
 
-    float discriminant = halfB * halfB - a * c;
-    if (discriminant < 0)
+    double discriminant = halfB * halfB - a * c;
+    if (discriminant < 0.0)
         return false;
-    float sqrtDiscriminant = sqrt(discriminant);
+    double sqrtDiscriminant = sqrt(discriminant);
 
-    float root = (-halfB - sqrtDiscriminant) / a;
+    double root = (-halfB - sqrtDiscriminant) / a;
     if (root < tMin || root > tMax) {
         root = (-halfB + sqrtDiscriminant) / a;
         if (root < tMin || root > tMax)
@@ -29,7 +29,7 @@ bool Sphere::Hit(const Ray &ray, float tMin, float tMax, HitRecord &hitRecord) c
 
     hitRecord.m_T = root;
     hitRecord.m_HitPoint = ray.GetPositionAtDistance(hitRecord.m_T);
-    glm::vec3 outwardNormal = (hitRecord.m_HitPoint - m_Center) / m_Radius;
+    glm::dvec3 outwardNormal = (hitRecord.m_HitPoint - m_Center) / m_Radius;
     hitRecord.m_SetFaceNormal(ray, outwardNormal);
 
     return true;
